Moving extractEnum up into SwaggerTerms
