+++ modules/codegen/src/main/scala/com/twilio/guardrail/terms/SwaggerTerms.scala
@@ -26,6 +26,7 @@ abstract class SwaggerTerms[L <: LA, F[_]] {
   def MonadF: Monad[F]
 
   def extractCommonRequestBodies(components: Tracker[Option[Components]]): F[Map[String, RequestBody]]
+  def extractEnum(swagger: Tracker[Schema[_]]): F[Either[String, List[String]]]
   def extractOperations(
       paths: Tracker[Mappish[List, String, PathItem]],
       commonRequestBodies: Map[String, RequestBody],
